setMitglied(Mitglied mitglied) in Spendenbescheiung rmi

+++ src/main/java/de/jost_net/JVerein/gui/control/SpendenbescheinigungControl.java
@@ -850,7 +850,7 @@ public class SpendenbescheinigungControl extends DruckMailControl
         }
         else
         {
-          spendenbescheinigung.setMitglied(null);
+          spendenbescheinigung.setMitglied((Integer) null);
         }
       }
       catch (Exception e)
